+++ system/src/documents/CompendiumsSD.mjs
@@ -187,7 +187,7 @@ export default class CompendiumsSD {
 	}
 
 	static async npcAttacks(sources=[]) {
-		return CompendiumsSD.talents(["NPC Attack"], sources);
+		return CompendiumsSD._documents("Item", "NPC Attack", sources);
 	}
 
 	static async npcFeatures(sources=[]) {
@@ -230,7 +230,7 @@ export default class CompendiumsSD {
 	}
 
 	static async scrolls(sources=[]) {
-		return CompendiumsSD.talents(["Scroll"], sources);
+		return CompendiumsSD._documents("Item", "Scroll", sources);
 	}
 
 	static async spellcastingClasses(sources=[]) {
@@ -251,7 +251,7 @@ export default class CompendiumsSD {
 	}
 
 	static async spells(sources=[]) {
-		return CompendiumsSD.talents(["Spell"], sources);
+		return CompendiumsSD._documents("Item", "Spell", sources);
 	}
 
 	static async talents(subtypes=[], sources=[]) {